api/requirements.txt
fastapi>=0.104.1
python-multipart>=0.0.6
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.2
chromadb>=0.4.18
sentence-transformers>=2.2.2
pydantic>=2.0.0
uvicorn>=0.24.0
